copr: build correct srpm and add master suffix
- Add a master/date suffix for COPR builds.
Only add the date without time, as two different packages need to be
build in COPR from this repository with exactly the same version number.
If we used time, it could be slightly different.

- Only build the relevant srpm package.

This package builds 2 rpm's:
- ovirt-engine-wildfly
- ovirt-engine-wildfly-overlay

To make sure the correct one is build, we pass the spec that needs to be
build via COPR and then call the correct build script.
Before this change both srpm's were build and then only the correct one
copied to the outdir, but this was broken.

Signed-off-by: Jean-Louis Dupond <jean-louis@dupond.be>